fully block inquiry call and easier clean up
