+++ src/applications/ask-va/containers/ResponseInboxPage.jsx
@@ -14,7 +14,6 @@ import PropTypes from 'prop-types';
 import React, { useCallback, useEffect, useState } from 'react';
 import { connect } from 'react-redux';
 import { withRouter } from 'react-router';
-import { useFeatureToggle } from 'platform/utilities/feature-toggles';
 import BreadCrumbs from '../components/BreadCrumbs';
 import FileUpload from '../components/FileUpload';
 import NeedHelpFooter from '../components/NeedHelpFooter';
@@ -55,18 +54,6 @@ const ResponseInboxPage = ({ router }) => {
     return pathArray[pathArray.length - 1];
   };
   const inquiryId = getLastSegment();
-
-  const {
-    TOGGLE_NAMES,
-    useToggleLoadingValue,
-    useToggleValue,
-  } = useFeatureToggle();
-
-  const toggleOldPortalAlert = TOGGLE_NAMES.askVaAlertLinkToOldPortal;
-  const isOldPortalAlertEnabled = useToggleValue(toggleOldPortalAlert);
-  const isLoadingFeatureFlags = useToggleLoadingValue(toggleOldPortalAlert);
-  const showAlertAndHideForm =
-    !isLoadingFeatureFlags && isOldPortalAlertEnabled;
 
   const postApiData = useCallback(
     async url => {
@@ -213,8 +200,7 @@ const ResponseInboxPage = ({ router }) => {
 
   useEffect(
     () => {
-      if (inquiryId && !showAlertAndHideForm)
-        getApiData(`${envUrl}${URL.GET_INQUIRIES}/${inquiryId}`);
+      if (inquiryId) getApiData(`${envUrl}${URL.GET_INQUIRIES}/${inquiryId}`);
     },
     [inquiryId, getApiData],
   );
@@ -245,26 +231,6 @@ const ResponseInboxPage = ({ router }) => {
     return (
       <VaAlert status="info" className="vads-u-margin-y--4">
         No inquiry data available.
-      </VaAlert>
-    );
-  }
-
-  if (showAlertAndHideForm) {
-    return (
-      <VaAlert
-        className="vads-u-margin-x--2 vads-u-margin-top--3  vads-u-margin-bottom--9"
-        close-btn-aria-label="Close notification"
-        status="warning"
-        uswds
-      >
-        <h2 id="track-your-status-on-mobile" slot="headline">
-          This version of Ask VA isn’t working right now
-        </h2>
-        <p className="vads-u-margin-y--0">
-          We’re making some updates to Ask VA on VA.gov. We’re sorry it’s not
-          working right now. You can still use the{' '}
-          <a href="https://ask.va.gov/">previous version of Ask VA</a>
-        </p>
       </VaAlert>
     );
   }
